# dnsproxy: Improve regex pattern