_TL;DR:_ This is now V3 of https://github.com/cilium/cilium/pull/20214.

The full list of changes can be found in the commits (tho. this is based off https://github.com/cilium/cilium/pull/21288);
- Extract regex anchors to the beginning and end of the full pattern
- Remove the use of regex groups since we don't use them
- Short circuit matchPatterns with just a wildcard and match all fqdns
- Skip running `.MatchString` on wildcard regexes.

The overall perf change for all commits in this PR vs. https://github.com/cilium/cilium/pull/21288 (all without the lru cache, and ignore the time/op, since they are running in docker for mac 😢 ):

```
$ benchstat <pre-this-pr> <this-pr>
name                             old time/op          new time/op          delta
_perEPAllow_setPortRulesForID-7           77.4s ± 8%           83.0s ±17%     ~     (p=0.310 n=5+5)

name                             old B(HeapInUse)/op  new B(HeapInUse)/op  delta
_perEPAllow_setPortRulesForID-7           97.3M ± 1%           74.2M ± 1%  -23.75%  (p=0.008 n=5+5)

name                             old alloc/op         new alloc/op         delta
_perEPAllow_setPortRulesForID-7          78.6GB ± 0%          73.0GB ± 0%   -7.18%  (p=0.008 n=5+5)

name                             old allocs/op        new allocs/op        delta
_perEPAllow_setPortRulesForID-7            167M ± 0%            108M ± 0%  -35.37%  (p=0.008 n=5+5)
```
---

After this we plan doing;
- Switch from string concatenation to a byte buffer in `GeneratePattern`
- Consider the regex for matchPatterns + `map[string]struct{}` for matchNames.
- Replacing the `[.]` with `\.` to reduce the regex size
- Replacing the `*` replacement from `[-a-zA-Z0-9_]*` to `[^.]*`
- ...

--

- [ ] For first time contributors, read [Submitting a pull request](https://docs.cilium.io/en/stable/contributing/development/contributing_guide/#submitting-a-pull-request)
- [ ] All code is covered by unit and/or runtime tests where feasible.
- [ ] All commits contain a well written commit description including a title,
      description and a `Fixes: #XXX` line if the commit addresses a particular
      GitHub issue.
- [ ] All commits are signed off. See the section [Developer’s Certificate of Origin](https://docs.cilium.io/en/stable/contributing/development/contributing_guide/#dev-coo)
- [ ] Provide a title or release-note blurb suitable for the release notes.
- [ ] Thanks for contributing!

<!-- Description of change -->

Fixes: #issue-number

```release-note
dnsproxy: Improve regex used for matching dns queries by reducing its complexity and size to save memory and speed up matching
```