@odinuge 2022-06-20
Hey @christarazi! Thanks! Did the initial testing on the v1.9 release branch, since thats what we are using atm., so didn't see those benchmarks.

I don't have a set of policies at hand atm., I've just been creating a set of domains to push through. Do you have the CNP list you were using in that PR?

Also, discussed this with a colleague, and we decided to try using a set (essentially a `map[string]bool`) for the fqdn names, and then keeping the regex for patterns. That was waaaaay faster on the synthetic tests as well, and the memory allocations aren't even comparable. Will need to test in some real environment first tho. Will hopefully do some quick implementation this week, and will ship my benchmarking code as soon as I can, hopefully before end of week.