@odinuge 2022-06-21
Did a quick test now with this PR vs. master, with a lru size of 128 with the same dataset as @christarazi was using with his initial testing;

```
# This PR
$ go test -tags privileged_tests -v -run '^$' -bench Benchmark_perEPAllow_setPortRulesForID_large -benchmem -benchtime 1x -memprofile memprofile.out ./pkg/fqdn/dnsproxy
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/fqdn/dnsproxy
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
Benchmark_perEPAllow_setPortRulesForID_large
Before (N=1)
Alloc = 32 MiB  HeapInuse = 53 MiB      Sys = 982 MiB   NumGC = 15
After (N=1)
Alloc = 369 MiB HeapInuse = 378 MiB     Sys = 982 MiB   NumGC = 37
Before (N=1)
Alloc = 34 MiB  HeapInuse = 52 MiB      Sys = 1017 MiB  NumGC = 48
After (N=1)
Alloc = 237 MiB HeapInuse = 300 MiB     Sys = 1017 MiB  NumGC = 71
Benchmark_perEPAllow_setPortRulesForID_large-7                 1        5772238001 ns/op        2161473136 B/op 23052832 allocs/op
PASS
ok      github.com/cilium/cilium/pkg/fqdn/dnsproxy      21.727s
```
```
# Baseline / master
root@eab730ccc555:/code# go test -tags privileged_tests -v -run '^$' -bench Benchmark_perEPAllow_setPortRulesForID_large -benchmem -benchtime 1x -memprofile memprofile.out ./pkg/fqdn/dnsproxy
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/fqdn/dnsproxy
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
Benchmark_perEPAllow_setPortRulesForID_large
Before (N=1)
Alloc = 32 MiB  HeapInuse = 50 MiB      Sys = 990 MiB   NumGC = 14
After (N=1)
Alloc = 464 MiB HeapInuse = 484 MiB     Sys = 990 MiB   NumGC = 35
Before (N=1)
Alloc = 36 MiB  HeapInuse = 54 MiB      Sys = 1101 MiB  NumGC = 45
After (N=1)
Alloc = 527 MiB HeapInuse = 538 MiB     Sys = 1101 MiB  NumGC = 66
Benchmark_perEPAllow_setPortRulesForID_large-7                 1        5331359995 ns/op        3019854640 B/op 26485069 allocs/op
PASS
ok      github.com/cilium/cilium/pkg/fqdn/dnsproxy      23.638s
```

We clearly see the heap usage is very much lower with this PR. Still need some tweaking tho. Will try to grab a bigger dataset and check the diff