@odinuge 2022-12-06
Hi, thanks!

> One thing I just realized is that once this lands, will it change the behavior of existing policies? Is there a potential it might break on existing matchpatterns in user environments?

This should be a noop for users, since what the regex will match is the same, its just the regex pattern that is different. It should also be fully backwards compatible with older versions with the restore feature as well (unless I'm missing anything obvious..).

I think the unit-tests [here](https://github.com/cilium/cilium/blob/master/pkg/fqdn/dnsproxy/proxy_test.go) and [here](https://github.com/cilium/cilium/blob/master/pkg/fqdn/dnsproxy/proxy_test.go) should cover all cases, and make sure that the end user behavior is the same.