@aanm 2023-01-30
Travis failed with https://github.com/cilium/cilium/issues/23314