@odinuge 2023-01-30
Thanks @aanm!

We deployed this to production here over the last few weeks, and on the hosts with the biggest l7 rulests and the most dns queries, including a `matchPattern: *` rule for some stuff, we saw a memory reduction of more than 50% (the graph is per host/node, and the drop is when deploying a version with this patchset).


<img width="1455" alt="Screen Shot 2023-01-30 at 11 52 51 AM" src="https://user-images.githubusercontent.com/1467188/215470092-99eb8688-45b0-45ed-9a17-0e4264ecfb18.png">